add optional embedding of collection items in view
Enable with optional boolean query parm "embed" (example: "embed=1" or "embed=true").
When embedding, the full resource respresentations are used by default.
However a new "fields" option can be specified with a comma-delimited
list of the fields to include. Doing so limits not only the
fields in the json view, but also in the mongodb query.
This is accomplished with a new "fields" optional parameter to the various
get_children methods in Collection.  This should only be used as a read-only
optimization.  Saving objects loaded this way could result in data loss.